refactor(engine_license): move license_policy helpers to infrastructure/helpers
