+++ tools/devsecops_engine_tools/engine_sca/engine_license/test/infrastructure/helpers/test_license_policy.py
@@ -1,4 +1,4 @@
-from devsecops_engine_tools.engine_sca.engine_license.src.domain.usecases.license_policy import (
+from devsecops_engine_tools.engine_sca.engine_license.src.infrastructure.helpers.license_policy import (
     build_policy_from_remote_config,
     classify_package,
     get_value,
